requirements.txt
# checkov
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiomultiprocess==0.9.1
aiosignal==1.3.1
annotated-types==0.7.0
argcomplete==3.5.1
attrs==24.2.0
bc-detect-secrets==1.5.22
bc-jsonpath-ng==1.6.1
bc-python-hcl2==0.4.2
beartype==0.19.0
beautifulsoup4==4.12.3
boolean.py==4.0
boto3==1.35.49
botocore==1.35.58
cached-property==2.0.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
checkov==3.2.289
click==8.1.7
click-option-group==0.5.6
cloudsplaining==0.7.0
colorama==0.4.6
ConfigArgParse==1.7
contextlib2==21.6.0
cyclonedx-python-lib==6.4.4
decorator==5.1.1
defusedxml==0.7.1
docker==7.1.0
dockerfile-parse==2.0.1
dpath==2.1.3
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib_metadata==7.2.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
junit-xml==1.9
lark==1.2.2
license-expression==30.4.0
Markdown==3.7
MarkupSafe==3.0.2
multidict==6.1.0
networkx==2.6.3
numpy==2.1.3
openai==0.28.1
orjson==3.10.11
packageurl-python==0.13.4
packaging==23.2
ply==3.11
policy_sentry==0.13.1
prettytable==3.12.0
propcache==0.2.0
py-serializable==1.1.2
pycares==4.4.0
pycep-parser==0.5.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rdflib==7.1.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.21.0
rustworkx==0.13.2
s3transfer==0.10.3
schema==0.7.5
semantic-version==2.10.0
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
soupsieve==2.6
spdx-tools==0.8.3
tabulate==0.9.0
termcolor==2.3.0
tqdm==4.67.0
typing_extensions==4.12.2
unidiff==0.7.5
uritools==4.0.3
urllib3==2.2.3
wcwidth==0.2.13
xmltodict==0.14.2
yarl==1.17.1
zipp==3.21.0

# PyGithub
Deprecated==1.2.14
PyGithub==2.5.0
cryptography==43.0.3
pyjwt==2.9.0
pynacl==1.5.0
wrapt==1.16.0